feat(TrustProviders): Remove callbacks from the trust provider sdk
The GetStatus and its handler never made sense in the context of an adapter. Callbacks don't get
called with tokens and it doesn't seem like something they should do. It just so happens that our
adapters are a convenient place to put the Callbacks, but that doesn't mean they are part of the
trust providers.